Day 6/MENTOR_AI/requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
pydantic
PyPDF2
beautifulsoup4 
request